共通鍵暗号方式での暗号化

+++ password_manager.sh
@@ -2,6 +2,9 @@
 
 echo "ようこそパスワードマネージャーへ！"
 
+read -s -p "パスフレーズを入力してください:" PASSPHRASE
+echo " "
+
 while true 
 do
 	read -p "次の選択肢から入力してください(Add Password/Get Password/Exit):" select
@@ -12,19 +15,46 @@ do
 	       echo "Thank you!"
 	       break
 
+
+
         #Add Passwordが入力された場合
         elif [ "${select}" = "Add Password" ]; then
 	       read -p "サービスを入力してください:" service
                read -p "ユーザー名を入力してください:" user
                read -s -p "パスワードを入力してください:" password
                echo " "
-               echo "service:${service},user:${user},password:${password}" >> password_manager.txt.pgp --batch --passphrase "password" --symmeric password_manager.txt
+
+	       #既存の暗号化ファイルがある場合に複合化する
+	       if [ -f "password_manager.txt.gpg" ];then
+		       gpg --batch --passphrase "${PASSPHRASE}" --decrypt password_manager.txt.gpg > password_manager.txt 2>/dev/null
+               fi
+
+	       #直前のコマンドが成功したか確認
+               if [ $? -ne 0 ];then
+		       echo "パスフレーズが違います。"
+		       exit 1
+	       fi
+
+               #データの追記
+               echo "service:${service},user:${user},password:${password}" >> password_manager.txt
+	       #再度暗号化して元のファイルを削除
+               gpg --batch --yes --passphrase "${PASSPHRASE}" --symmetric -o password_manager.txt.gpg password_manager.txt 2>/dev/null
 	       rm password_manager.txt
                echo "パスワードの追加に成功しました"
+
+
+
 	#Get Passwordが入力された場合
          elif [ "${select}" = "Get Password" ]; then
                read -p "サービスを入力してください:" service
-               result=$(gpg --batch --passphrase "password" --decrypt password_manager.txt.pgp 2>/dev/null | grep "service:${service}")
+               result=$(gpg --batch --passphrase "${PASSPHRASE}" --decrypt password_manager.txt.gpg 2>/dev/null | grep "service:${service}")
+               
+               #パスフレーズの確認
+	       if [ $? -ne 0 ];then
+		       echo "パスフレーズが違います。"
+		       exit 1
+	       fi
+
 	       if [ -z "${result}" ]; then
 		       echo "そのサービスは登録されていません。"
 	       else
